Crate strymon_model

Shared data types for Strymon Core.

This crate defines a number of common data types used by the various components of Strymon Core. Most types in this crate are used to store and the metadata of various enities in the system, such as running jobs, available executors, published topics and so forth.

All types in the root module can be serialized both with Serde and Abomonation. The data types used in the root module of this crate are all part of the catalog and inform users about the current state of the system.

The types found in the config module on the other hand are purely implementation details and define the format in which configuration data is passed between the different components.

Modules

config
Data-types used to represent the configuration of the various Strymon components.

Structs

Executor
The meta-data of an available executor in the system.

ExecutorId
A unique numerical identifier of an executor instance.

Job
The meta-data of a submitted and running job.

JobId
A unique numerical identifier for a job.

JobProgram
The meta-data about the executable code of a job.

Publication
Associates the publication of a topic by its publishing job.

Subscription
Associates the subscription to a topic by a subscribing job.

Topic
The metadata of a published topic.

TopicId
A unique numerical identifier for a topic.

TopicType
A representation of the type of data found in a topic.

Enums

ExecutionFormat
The format of a job executable.

TopicSchema
The kind of protocol used in a topic.