Enum strymon_model::config::job::EnvError

pub enum EnvError {
    VarErr(VarError),
    IntErr(ParseIntError),
}

Error which occurs when the job process configuration cannot be parsed from the local environment.

Variants
VarErr(VarError)

The configuration data could not be parsed because of a missing environment variable.

IntErr(ParseIntError)

The configuration data contained invalid numbers.